% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLR_activate.R
\name{getLR_activate}
\alias{getLR_activate}
\title{Identify the methods in which the Ligand-Receptor is activated.}
\usage{
getLR_activate(CCC_net, LR_pairs)
}
\arguments{
\item{CCC_net}{Cell-cell communication network.}

\item{LR_pairs}{A String of ligand-receptor connected by “_”.}
}
\description{
Identify the methods in which the Ligand-Receptor is activated.
}
