% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOncoLR_interfaces.R
\name{getOncoLR_interfaces}
\alias{getOncoLR_interfaces}
\title{Show which oncoLR occurs interfaces.}
\usage{
getOncoLR_interfaces(OncoLR, mutation_data)
}
\arguments{
\item{OncoLR}{Preferred oncoLR.}

\item{mutation_data}{Mutation data.}
}
\description{
Show which oncoLR occurs interfaces.
}
