% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOncoLRI.R
\name{getOncoLRI}
\alias{getOncoLRI}
\title{Preferred ligand-receptor pairs.}
\usage{
getOncoLRI(CCC_net, exp1, exp2, mutation_data, q = 0.01)
}
\arguments{
\item{CCC_net}{Cell-cell communication network.}

\item{exp1}{Cancer Gene Expression Profiling File 1, FPKM Values.}

\item{exp2}{Cancer gene expression profile file 2, FPKM_UQ values.}

\item{mutation_data}{a dataframe, defaults to TCGA-MC3 project data.}

\item{q}{The preference threshold, which determines whether the ligand-receptor pair is significant, has a default value of 0.01.}
}
\description{
Preferred ligand-receptor pairs.
}
