% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTumorLR.R
\name{getTumorLR}
\alias{getTumorLR}
\title{Organizing CCC results for single-cell data and multiple idle slices based on the dplyr package.}
\usage{
getTumorLR(SC_net, ST_net, is.union = T, is.intersect = F)
}
\arguments{
\item{SC_net}{CCC results inferred from single-cell data.}

\item{ST_net}{CCC results inferred from spatial transcriptome data.}

\item{is.union}{Default TRUE. Take the union set of the CCC results inferred for the slices.}

\item{is.intersect}{Default FALSE，if TRUE，Take the intersect set of the CCC results inferred for the slices.}
}
\description{
Organizing CCC results for single-cell data and multiple idle slices based on the dplyr package.
}
