% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTumorLR_SC.R
\name{getTumorLR_SC}
\alias{getTumorLR_SC}
\title{Constructing a cell-cell interaction network involving tumor cells from the single-cell level.}
\usage{
getTumorLR_SC(sc_data, sc_count, sc_metadata, ligand_receptor_DB = NULL)
}
\arguments{
\item{sc_data}{Seurat object to create ligand receptor matrices from.}

\item{sc_count}{Pre-processed (i.e. no raw counts) gene expression matrix with genes on rows and cells on columns; rownames(gene_expr) must contain the genes names.}

\item{sc_metadata}{Grouping variable in the Seurat object metadata used to define the groups of cells. Defaults to two columns, cell_id and cell_types containing the ‘Tumor’ cell type.}

\item{ligand_receptor_DB}{Data.frame of ligands, receptors and interactions，The default is the TICC and Omnipath datasets. If NULL, database of ligand/receptors interactions that comes with the method is used.}
}
\description{
Constructing a cell-cell interaction network involving tumor cells from the single-cell level.
}
