% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTumorLR_ST.R
\name{getTumorLR_ST}
\alias{getTumorLR_ST}
\title{Constructing a cell-cell interaction network involving tumor cells from the spatial transcriptome level.}
\usage{
getTumorLR_ST(
  st_count,
  st_metadata,
  ligand_receptor_DB,
  python_path,
  outputDir
)
}
\arguments{
\item{st_count}{A dataframe containing spatial transcriptome counts with each column representing a cell and each row representing a gene.}

\item{st_metadata}{A dataframe containing spatial transcriptomics coordinates with 4 columns, i.e., cell id, x, y indicating single cell spatial transcriptomics data and cell_types containing the ‘Tumor’ cell type.}

\item{ligand_receptor_DB}{Data.frame of ligands, receptors and interactions.}

\item{python_path}{path to python binary to use}

\item{outputDir}{Character, The output path of the currently running job where temporary and final results will be saved.}
}
\description{
Constructing a cell-cell interaction network involving tumor cells from the spatial transcriptome level.
}
