% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.Plot.R
\name{lollipop.Plot}
\alias{lollipop.Plot}
\title{Draws lollipop plot of interface residues on  Protein structure.}
\usage{
lollipop.Plot(
  gene_mutation,
  gene_residues_data,
  protein.df,
  protein_height = 0.1,
  protein_fill = NA,
  protein_color = "#6b8e23",
  protein_size = 1,
  str.fill = "#cdbde2",
  str.col = "#cdbde2",
  str.size = 0,
  residule.fill = "#FF0000",
  residule.col = "#FF0000",
  residule.size = 0,
  line.col = "grey50",
  residule_height = 0.5,
  linewidth = 0.5,
  point_shape = 21,
  point_cols,
  point_fill,
  point_stroke = 0.2,
  point_size = 3,
  is.lable = F,
  shift.factor = 0.02
)
}
\arguments{
\item{gene_mutation}{Gene mutation data, including 4 columns, the first column AA indicates the gene amino acid site, the second column Mut is set to HGVSp_Short, the third column Type defaults to the mutation type, and the fourth column is the gene mutation count.}

\item{gene_residues_data}{Gene interface data, including 4 columns.}

\item{protein.df}{Protein structure length data.}

\item{protein_height}{map parameter}

\item{protein_fill}{map parameter}

\item{protein_color}{map parameter}

\item{protein_size}{map parameter}

\item{str.fill}{map parameter}

\item{str.col}{map parameter}

\item{str.size}{map parameter}

\item{residule.fill}{map parameter}

\item{residule.col}{map parameter}

\item{residule.size}{map paramete}

\item{line.col}{map parameter}

\item{residule_height}{map parameter}

\item{linewidth}{map parameter}

\item{point_shape}{map parameter}

\item{point_cols}{map parameter}

\item{point_fill}{map parameter}

\item{point_stroke}{map parameter}

\item{point_size}{map parameter}

\item{is.lable}{Logical value, whether to add a label}

\item{shift.factor}{map paramete}
}
\description{
Draws lollipop plot of interface residues on  Protein structure.
}
