% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muteffect.R
\name{muteffect}
\alias{muteffect}
\title{Calculate the effect of mutations on ligand/receptor functions.}
\usage{
muteffect(
  obj,
  LR_data,
  plot1 = T,
  plot2 = T,
  y_position1 = NULL,
  ylim1 = NULL,
  y_position2 = NULL,
  ylim2 = NULL
)
}
\arguments{
\item{obj}{Defaults to ANNOVAR annotated results, the result file is a csv file.}

\item{LR_data}{A dataset with at least ligand/receptor columns of data. The default is the TICC and Omnipath datasets.}

\item{plot1}{Default TRUE.}

\item{plot2}{Default TRUE.}

\item{y_position1}{numeric vector with the y positions of the brackets}

\item{ylim1}{Limits for the y axes.}

\item{y_position2}{numeric vector with the y positions of the brackets}

\item{ylim2}{Limits for the y axes.}
}
\description{
Calculate the effect of mutations on ligand/receptor functions.
}
