% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutfreq.R
\name{mutfreq}
\alias{mutfreq}
\title{Calculate mutation frequency}
\usage{
mutfreq(mutation_obj)
}
\arguments{
\item{mutation_obj}{a dataframe generated from the TCGA-MC3 project data, the user-defined dataset need to include at least Hugo_Symbol, Tumor_Sample_Barcode, Variant_Classification columns.}
}
\value{
mutation_obj_frequency
}
\description{
Calculate mutation frequency
}
