% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutsummary.R
\name{mutsummary}
\alias{mutsummary}
\title{The mutation summary}
\usage{
mutsummary(mutation.obj, plotmafSummary, plotTiTv)
}
\arguments{
\item{mutation.obj}{an MAF object generated by read.maf, defaults to TCGA-MC3 project data, the user-defined dataset need to include at least Hugo_Symbol, Tumor_Sample_Barcode, Variant_Classification columns.}

\item{plotmafSummary}{logical, defaults to TRUE, plots maf summary.}

\item{plotTiTv}{logical, defaults to TRUE，plot Transition and Trasnversion ratios. Takes results generated from titv and plots the Ti/Tv ratios and contributions of 6 mutational conversion classes in each sample.}
}
\description{
Map mutation information based on the maftools package.
}
