% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoLR_Target_enrichment.R
\name{OncoLR_Target_enrichment}
\alias{OncoLR_Target_enrichment}
\title{Enrichment analysis of LR inferred target genes based on clusterProfiler package.}
\usage{
OncoLR_Target_enrichment(gene, isRunGO = T, isRunKEGG = T)
}
\arguments{
\item{gene}{a vector of gene.}

\item{isRunGO}{Default TRUE. Run GO enrichment analysis.}

\item{isRunKEGG}{Default TRUE. Run KEGG enrichment analysis.}
}
\description{
Enrichment analysis of LR inferred target genes based on clusterProfiler package.
}
