% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoLR_WM_cor.R
\name{OncoLR_WM_cor}
\alias{OncoLR_WM_cor}
\title{Characterizing the association of oncoLRI in mutation and wild-type.}
\usage{
OncoLR_WM_cor(
  OncoLR,
  exp,
  mutation_data,
  ligand.source.cell = "Tumor cell",
  receptor.source.cell = "Other cell"
)
}
\arguments{
\item{exp}{Cancer Gene Expression Profiling File, FPKM Values.}

\item{mutation_data}{A dataframe, defaults to TCGA-MC3 project data.}

\item{ligand.source.cell}{Ligand source cells. Default 'Tumor cell'}

\item{receptor.source.cell}{Receptor source cells. Default 'Other cell'}

\item{OncoLRI}{Preferred oncoLRI.}
}
\description{
Characterizing the association of oncoLRI in mutation and wild-type.
}
