% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunOncoLR_Target_net.R
\name{RunOncoLR_Target_net}
\alias{RunOncoLR_Target_net}
\title{Constructs cell-to-cell post-preference Ligand-Receptor-TarGene networks based on scRNA-Seq expression matrices and cell_type tables}
\usage{
RunOncoLR_Target_net(
  sc_data,
  cell_type,
  oncoLR,
  ligand.source.cell = "Tumor cell",
  receptor.source.cell = "Other cell",
  ligand_target_matrix,
  weighted_networks
)
}
\arguments{
\item{sc_data}{A Seurat object}

\item{cell_type}{A character vector representing ligand/recepor cell types.}

\item{oncoLR}{A dataframe, including ligand, receptor, interaction_name.}

\item{ligand.source.cell}{Ligand source cells. Default 'Tumor cell'}

\item{receptor.source.cell}{Receptor source cells. Default 'Other cell'}

\item{ligand_target_matrix}{Indicates the regulation score between ligand-target (column is ligand). The default setting is the file “ligand_target_matrix_nsga2r_final.rds” in the /databases/ folder.}

\item{weighted_networks}{Weighted network representing interactions and their weights in the ligand signaling and gene regulatory networks. The default setting is the file “weighted_networks_nsga2r_final.rds” in the /databases/ folder.}
}
\description{
Constructs cell-to-cell post-preference Ligand-Receptor-TarGene networks based on scRNA-Seq expression matrices and cell_type tables
}
