% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_Lig_Rec_TGs.R
\name{Select_Lig_Rec_TGs}
\alias{Select_Lig_Rec_TGs}
\title{Select_Lig_Rec_TGs}
\usage{
Select_Lig_Rec_TGs(
  ExprMat,
  AnnoMat,
  LocaMat,
  Databases,
  python_path,
  min.pct,
  expr.ct,
  pct.ct
)
}
\arguments{
\item{ExprMat}{A dataframe containing spatial transcriptome counts with each column representing a cell and each row representing a gene.}

\item{AnnoMat}{A dataframe containing spatial transcriptomics coordinates with 4 columns, i.e., cell id, x, y indicating single cell spatial transcriptomics data and cell_types containing the ‘Tumor’ cell type.}

\item{LocaMat}{spot_loc}

\item{Databases}{path to python binary to use}

\item{pct.ct}{}
}
\description{
Select_Lig_Rec_TGs
}
