% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TumorLR_multiSlide.R
\name{TumorLR_multiSlide}
\alias{TumorLR_multiSlide}
\title{Organizing CCC results for multiple idle slices based on the dplyr package.}
\usage{
TumorLR_multiSlide(net1, net2, is.union = T, is.intersect = F)
}
\arguments{
\item{net1}{CCC results for slice 1 based on inferences from the getTumorLR_ST() function.}

\item{net2}{CCC results for slice 2 based on inferences from the getTumorLR_ST() function.}

\item{is.union}{Default TRUE. Take the union set of the CCC results inferred for the slices.}

\item{is.intersect}{Default FALSE，if TRUE，Take the intersect set of the CCC results inferred for the slices.}
}
\description{
Organizing CCC results for multiple idle slices based on the dplyr package.
}
